financial-analysis/requirements.txt
# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Machine Learning
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# LLM APIs
openai>=1.0.0
anthropic>=0.18.0
tiktoken>=0.5.0

# Web & API
fastapi>=0.109.0
uvicorn>=0.27.0
streamlit>=1.30.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.26.0
tenacity>=8.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Development
black>=24.0.0
ruff>=0.1.0
